feature: rename activity bar menu entries to activity bar quickpick menu entries
