+++ packages/renderer-worker/src/parts/ViewletActivityBar/ViewletActivityBar.ipc.js
@@ -10,7 +10,7 @@ export * from './ViewletActivityBar.js'
 export * from './ViewletActivityBarCommands.js'
 export * from './ViewletActivityBarCss.js'
 export * from './ViewletActivityBarKeyBindings.js'
-export * from './ViewletActivityBarMenuEntries.js'
+export * from './ViewletActivityBarQuickPickMenuEntries.js'
 export * from './ViewletActivityBarName.js'
 export * from './ViewletActivityBarRenderVirtualDom.js'
 export * from './ViewletActivityBarResize.js'
