+++ packages/renderer-worker/src/parts/ViewletActivityBar/ViewletActivityBarQuickPickMenuEntries.js
